try a 'death' workaround -- cannot reproduce locally
only on travis-ci. presumably because it has load issues. so what we'll
do, is even though we've just called select(2), with nothing to read,
and waitpid(2) tells us the child is dead -- select(2) *again*, for any
more data. it would then return, and the *next* call, ehcking
waitpid(2), would indicate EOF. Otherwise, still the 'Very slow
platform' EOF is raised.