Test and document PC_MAX_CANON
Speak about canonical mode in write() family of functions,
and assert the result of IMAXBEL in test cases by using
stty(1).

There are inconsistent results using Travis-CI, so it is
disabled there, but tests are confirmed using OSX, Linux,
and OpenSolaris, and their respective limits documented
in the send() method.